+++ CHANGELOG.md
@@ -7,12 +7,41 @@
 
 ### Dependencies
 
-- Bump Android SDK from v6.13.0 to v6.13.1 ([#437](https://github.com/getsentry/sentry-android-gradle-plugin/pull/437))
-  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6131)
-  - [diff](https://github.com/getsentry/sentry-java/compare/6.13.0...6.13.1)
-- Bump CLI from v2.12.0 to v2.13.0 ([#439](https://github.com/getsentry/sentry-android-gradle-plugin/pull/439))
-  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2130)
-  - [diff](https://github.com/getsentry/sentry-cli/compare/2.12.0...2.13.0)
+- Bump CLI from v2.16.1 to v2.17.2 ([#469](https://github.com/getsentry/sentry-android-gradle-plugin/pull/469))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2172)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.16.1...2.17.2)
+
+## 3.5.0
+
+### Features
+
+- Add support for Logcat ([#455](https://github.com/getsentry/sentry-android-gradle-plugin/pull/455))
+- Write enabled instrumentations to manifest ([#441](https://github.com/getsentry/sentry-android-gradle-plugin/pull/441))
+
+### Dependencies
+
+- Bump CLI from v2.15.2 to v2.16.1 ([#458](https://github.com/getsentry/sentry-android-gradle-plugin/pull/458))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2161)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.15.2...2.16.1)
+- Bump Android SDK from v6.16.0 to v6.17.0 ([#462](https://github.com/getsentry/sentry-android-gradle-plugin/pull/462))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6170)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.16.0...6.17.0)
+
+## 3.4.3
+
+### Fixes
+
+- Make plugin Gradle 8 compatible ([#428](https://github.com/getsentry/sentry-android-gradle-plugin/pull/428))
+- Fix `uploadSentryNativeSymbols` task for Gradle 8 ([#447](https://github.com/getsentry/sentry-android-gradle-plugin/pull/447))
+
+### Dependencies
+
+- Bump Android SDK from v6.13.0 to v6.16.0 ([#442](https://github.com/getsentry/sentry-android-gradle-plugin/pull/442), [#449](https://github.com/getsentry/sentry-android-gradle-plugin/pull/449))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6160)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.13.0...6.16.0)
+- Bump CLI from v2.12.0 to v2.15.2 ([#439](https://github.com/getsentry/sentry-android-gradle-plugin/pull/439), [#443](https://github.com/getsentry/sentry-android-gradle-plugin/pull/443), [#446](https://github.com/getsentry/sentry-android-gradle-plugin/pull/446), [#450](https://github.com/getsentry/sentry-android-gradle-plugin/pull/450))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2152)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.12.0...2.15.2)
 
 ## 3.4.2